Provide default column order to DataTables in Dynamic Risk Report.
With DataTables being used for the Dynamic Risk Report, the default selected column was not being used for sorting, due to the default order POST parameter provided by DataTables being [0, 'asc']. Thus the SQL query was using the sort POST parameter to order its results, but then those results are reordered by the 0th column (usually ID) because of the default from DataTables.

This commit ensures the selected sort by column is used as the default for the table, and uses a sane default ordering for the respective columns. It still allows the user to change the ordering by clicking on the column header.