Change ordering of ID and risks to use numeric sorting.
The ID and risk values are currently being sorted using a string comparison function. This results in (for example) 2 being greater than 10.

This commit changes the logic behind searching the numeric columns id, calculated_risk, and residual_risk to be a proper numeric sort, handling either double-precision floating point (with an epsilon of 1e-4) or integer accuracy.