Validate upload filesize is smaller than or equal to PHP limits
PHP uses two variables that control the size of files that can be uploaded in a POST request, upload_max_filesize and post_max_size. If these are smaller than the SimpleRisk filesize limit and a file is uploaded that is below the SimpleRisk limit but greater than either of the above, then SimpleRisk will display an error to the end-user about a failed CRSF check, but the Apache logs will correctly record that one of the limits was exceeded.

This change displays a warning to an admin user when they change the upload limit if it exceeds either of those PHP limits, so they know what they need to change.

TODO: do localisation for other languages other than en.