Display a single value for equal min and max value ranges.
If the minimum and maximum values are the same for a given value range, this commit only displays one of the values with a currency prefix (arbitrarily choosing the minimum, but it shouldn't matter either way since they are the same), rather than the normal 'min to max'.

Our particular use-case for this feature is assets and mitigations that have not been valued are currently displayed as being in the lowest bracket, so we have set the lowest bracket to a minimum and maximum value of 0 to denote this fact. However, this is being rendered as '$0 to $0', which is not desirable. An alternative solution would be to allow asset values and mitigation costs to be null.