Display enabled/disabled status when editing a user
When an administrator is editing a user, they cannot easily see whether
that user is disabled or not, which can lead to confusion when debugging
user access to the system. This commit adds a new row to the table to
display this information.

It may also be desirable to make the enabled/disabled status toggleable
in this view, but I was not sure about the preferred interaction flow
for such a change, so have not done so.